% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlandsat-package.R
\docType{package}
\name{getlandsat-package}
\alias{getlandsat}
\alias{getlandsat-package}
\title{getlandsat - get Landsat 8 data from AWS public data sets}
\description{
getlandsat - get Landsat 8 data from AWS public data sets
}
\keyword{package}

