% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{lsat_cache}
\alias{lsat_cache}
\alias{lsat_cache_delete}
\alias{lsat_cache_delete_all}
\alias{lsat_cache_details}
\alias{lsat_cache_list}
\title{Manage cached files}
\usage{
lsat_cache_list(cache_path = "~/.landsat-pds")

lsat_cache_delete(files, cache_path = "~/.landsat-pds", force = FALSE)

lsat_cache_delete_all(cache_path = "~/.landsat-pds", force = FALSE)

lsat_cache_details(files = NULL, cache_path = "~/.landsat-pds")
}
\arguments{
\item{cache_path}{path to cached files. Default: \code{"~/.landsat-pds"}}

\item{files}{File name or names}

\item{force}{(logical) Should files be force deleted? Default: \code{FALSE}}
}
\description{
Manage cached files
}
\details{
\code{cache_delete} only accepts 1 file name, while \code{cache_delete_all}
doesn't accept any names, but deletes all files. For deleting many specific files,
use \code{cache_delete} in a \code{\link{lapply}} type call
}
\examples{
\dontrun{
# list files in cache
lsat_cache_list()

# List info for single files
lsat_cache_details(files = lsat_cache_list()[1])
lsat_cache_details(files = lsat_cache_list()[2])

# List info for all files
lsat_cache_details()

# delete files by name in cache
# lsat_cache_delete(files = lsat_cache_list()[1])

# delete all files in cache
# lsat_cache_delete_all()
}
}

