% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_list.R
\name{lsat_list}
\alias{lsat_list}
\title{List Landsat images}
\usage{
lsat_list(max = NULL, ...)
}
\arguments{
\item{max}{Integer indicating the maximum number of keys to return (max 1000,
default 1000).}

\item{...}{Further args passed on to \code{\link[aws.s3]{getbucket}}}
}
\description{
This is an alternative to using lsat_scenes(). This
function uses the AWS S3 API, while the other fxn simply downloads
the up to date compressed csv file.
}
\examples{
\dontrun{
lsat_list(max = 10)

# paging, start a specific key string
res <- lsat_list(max = 10)
lsat_list(marker = res$Key[10], max = 10)

# curl options
library("httr")
lsat_list(config = verbose())
}
}

