% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsat_scenes_files.R
\name{lsat_scene_files}
\alias{lsat_scene_files}
\title{List files for a Landsat scene}
\usage{
lsat_scene_files(x, ...)
}
\arguments{
\item{x}{(character) A URL to a scene html file}

\item{...}{Further args passed on to \code{\link[aws.s3]{getobject}}}
}
\description{
List files for a Landsat scene
}
\examples{
\dontrun{
res <- lsat_scenes(n_max = 10)
lsat_scene_files(x = res$download_url[1])
lsat_scene_files(x = res$download_url[2])
}
}

