% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsat_image.R
\name{lsat_image}
\alias{lsat_image}
\title{GET Landsat image(s)}
\usage{
lsat_image(x, path = NULL, overwrite = FALSE, ...)
}
\arguments{
\item{x}{(character) A file name for a geotif file, will be more general soon.}

\item{path}{(character) Path to content to.}

\item{overwrite}{(logical) Will only overwrite existing path if \code{TRUE}}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}}
}
\description{
GET Landsat image(s)
}
\examples{
\dontrun{
# pass an image name
lsat_image(x = "LC80101172015002LGN00_B5.TIF")
}
}

