% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsat_scenes.R
\name{lsat_scenes}
\alias{lsat_scenes}
\title{List Landsat scenes}
\usage{
lsat_scenes(extent = NULL, ...)
}
\arguments{
\item{extent}{extent vector, of the form min lon, max lon, min lat, max lat.
optional. requires \code{rgeos} and \code{sp} packages}

\item{...}{Further args passed on to \code{\link[readr:read_csv]{readr::read_csv()}}}
}
\description{
List Landsat scenes
}
\details{
We use \code{\link[readr:read_csv]{readr::read_csv()}} to read the scene list file
from \url{http://landsat-pds.s3.amazonaws.com/scene_list.gz}. See the help
file for \code{\link[readr:read_csv]{readr::read_csv()}} for what parameter you can pass to modify it's
behavior.

After reading the data from the web, we cache the file to disk, but remove
at the end of the R session.

This is an alternative to using \code{\link[=lsat_list]{lsat_list()}}. This function
downloads the up to date compressed csv file, while \code{\link[=lsat_list]{lsat_list()}}
uses the AWS S3 API.
}
\examples{
\dontrun{
res <- lsat_scenes()
res

# read only N rows
lsat_scenes(n_max = 10)

# query by extent
lsat_scenes(extent = c(-97, -95, 25, 30))
}
}
\seealso{
\code{\link[=lsat_scene_files]{lsat_scene_files()}}
}
