% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lsat_scenes.R
\name{lsat_scenes}
\alias{lsat_scenes}
\title{List Landsat scenes}
\usage{
lsat_scenes(...)
}
\arguments{
\item{...}{Further args passed on to \code{\link[readr]{read_csv}}}
}
\description{
List Landsat scenes
}
\examples{
\dontrun{
res <- lsat_scenes()
head(res)

# read only N rows
lsat_scenes(n_max = 10)
}
}

